refactor: use pydantic for screenshot manifest validation
