+++ pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
     "moderngl>=5.12.0",
     "numpy==2.*",
     "pillow>=11.2.1",
+    "pydantic>=2.11.5",
     'pyobjc-framework-Cocoa; platform_system == "Darwin"',
     "pyrr>=0.10.3",
     "scipy>=1.16.1",
